Do not cache incomplete completion candidates.
According to the protocol, the server can set the "isIncomplete" field to
indicate whether incremental completion should send requests to the server.
We cache the completion candidates as buffer local variables to store and
retrieve this info for the 'no-cache company command.

A new value of 'auto for `company-lsp-cache-candidates' customization is added.