Fix re-triggering condition.
The current logic checks whether (point) changes before and after post
completion actions to determine if re-triggering completion is desired. However,
this logic is not robust enough. It incorrectly re-triggers completion when the
post-completion action adds import statements by Javacomp, or fixes the
duplicated # characters by cquery (see #21).

This change proposes only re-trigger completion when the completed text before
point is any of the trigger characters. An additional custom variable is added
allowing users to turn this feature on/off.

I tested that it still works for the desired case (std::) while not
re-triggering completion for Javacomp auto-import case. However, it still
re-triggers completion for the cquery include case because `>' is one of the
trigger characters. Due to this reason, this feature is turned off by default.