Fixed several issues related to company-lsp--post-completion
1. Fixed issue related to expanding snippets after additionalEdits have been
applied. The snipped was expanded on wrong position becase applying
additionalEdits may move the points.

2. Fixed applying text edits when the completion has been triggered by completion
character. The old code was keeping the text in company-lsp--line-backup but if
you trigger the completion with . it was not part of company-lsp--line-backup.
As a result company-lsp was deleting the last character you have written.
Now the region to delete is calulated based on the inserted candidate which
works similar to company--insert-candidate.

3. When there is textEdit, the insert text should be the newText from the
textEdit.