Use the text from point to bol for prefix when caching
When the completion item contains a trigger completion char company-lsp
miscalulates the prefix. For example, if the user has typed `java.util.` and
completes "ArrayList" the company will calculate the prefix to be `util` but the
actual prefix is `java.util`. This change will fix the logic that reverts the
buffer state to be the same as the state at the time the completions were
requested.

- Fixes https://github.com/emacs-lsp/lsp-java/issues/95 and https://github.com/emacs-lsp/lsp-java/issues/46