Add options to filter candidate returned from server.
A new custom option `company-lsp-filter-candidates` is added to allow
customizing whether or not to filter candidates when
`company-lsp-cache-candidates` is set to nil. The option can be either set to a
boolean, or an alist for configuring it on a per-server basis.

When `company-lsp-cache-candidates` is set to non-nil values, this option
doesn't affect the filtering behavior since candidates are always filtered in
this case.

This should fix #79.