Fix range of snippet expansion with textEdit.
When expanding snippets, company-lsp calls `yas-expand-snippet` which replaces
text within given range with the insert text. Currently, the start of the range
is set to the start of the prefix. However, in some cases the start of text edit
range can be different from the start of the prefix. For example in
lsp-java (JDTLS), when completing `java.util`, the start of prefix is before
`util`. However the candidate can be `java.util.List`, whose start of range is
before `java`. Because currently we are using the prefix start, `java.util` will
become `java.java.util.List`, which is caused by replacing `util` with
`java.util.List`.

This change should fix emacs-lsp/lsp-java#67 and emacs-lsp/lsp-java#46